Style Changes and Markdown
There were a mix of underscore and camelCase function and variable names,
so all of those are now separated by underscores. More importantly, this
is the first test of including markdown syntax support to the text input.
All future posts will be written in markdown, and the after the titles are
parsed, the post body's will be converted to HTML through the python-markdown
module.