Grab next two posts when user scrolls to page bottom
When the user scrolls to the bottom of the page, instead of clicking the 'Moar
Posts' button, will simply be given the next two posts. This also removes the
behavior the auto scrolls the user to the top of the new posts, as the new
posts may load before the reader is done reading the current post.