Check to see if there are any more posts to get
If the user scrolls to the bottom of the page when there are no more posts
left, they would still be requested, leading to excess AJAX requests and
Javascript exceptions. The scroll event callback now checks to see if the 'Moar
Posts' button still exists before requesting more posts. The 'Moar Posts'
button is automatically removed on the previous request if there are no more
posts to get.